refactor: 496534 - Replace send-application-pack with send-application-pack-2
