+++ app/schema/portal/cdo/tasks/generic-task.js
@@ -4,7 +4,6 @@ const cdoTasksGetSchema = Joi.object({
   dogIndex: Joi.string().required(),
   taskName: Joi.string().valid(
     'send-application-pack',
-    'send-application-pack-2',
     'email-application-pack',
     'post-application-pack',
     'process-application-pack',
