+++ app/schema/portal/cdo/tasks/send-application-pack.js
@@ -1,11 +1,18 @@
 const Joi = require('joi')
 const { validatePayloadBuilder } = require('../../../../schema/common/validatePayload')
+const { errorMessages } = require('../../../error-messages')
 
 const schema = Joi.object({
   taskName: Joi.string().required(),
-  taskDone: Joi.string().messages({
-    '*': 'Confirm you\'ve sent the application pack'
-  }).required()
+  contact: Joi.string().messages({
+    '*': 'Select how you want to send the application pack'
+  }).required(),
+  email: Joi.alternatives().conditional('contact', {
+    is: 'email',
+    then: Joi.string().email().required(),
+    otherwise: Joi.any()
+  }).messages(errorMessages.email),
+  updateEmail: Joi.boolean().default(false)
 })
 
 const validateSendApplicationPack = (payload) => {
@@ -13,5 +20,6 @@ const validateSendApplicationPack = (payload) => {
 }
 
 module.exports = {
+  sendApplicationPackPayloadSchema: schema,
   validateSendApplicationPack
 }
