--- app/views/cdo/manage/tasks/send-application-pack-2.njk
@@ -1,62 +1,0 @@
-{% from "govuk/components/back-link/macro.njk" import govukBackLink %}
-{% from "govuk/components/input/macro.njk" import govukInput %}
-{% from "govuk/components/fieldset/macro.njk" import govukFieldset %}
-{% from "govuk/components/error-summary/macro.njk" import govukErrorSummary %}
-{% from "govuk/components/button/macro.njk" import govukButton %}
-{% from "govuk/components/radios/macro.njk" import govukRadios %}
-
-{% extends '_layout.njk' %}
-
-{% block pageTitle %}Send application pack - {{ serviceName }}{% endblock %}
-
-{% block beforeContent %}
-  <div class="govuk-grid-row">
-    <div class="govuk-grid-column-one-half">
-      {{ govukBackLink({
-        text: "Back",
-        href: model.backLink,
-        attributes: {
-          id: "back"
-        }
-        }) 
-      }}
-    </div>
-  </div>
-{% endblock %}
-
-{% block content %}
-<div class="govuk-grid-row">
-  <div class="govuk-grid-column-two-thirds">
-    {% if model.errors | length > 0 %}
-      {{ govukErrorSummary({
-        titleText: "There is a problem",
-        errorList: model.errors
-      }) }}
-    {% endif %}
-
-    <form method="POST" onsubmit="confirmSubmit.disabled = true; return true;" novalidate autocomplete="off">
-      <span class="govuk-caption-m">Dog {{ model.indexNumber }}</span>
-      <fieldset class="govuk-fieldset" {% if model.disabled == false %} aria-describedby="application-pack-hint"{% endif %}>
-        <legend class="govuk-fieldset__legend govuk-fieldset__legend--l govuk-!-margin-bottom-5">
-          <h1 class="govuk-fieldset__heading">How do you want to send the application pack?</h1>
-        </legend>
-        <div id="contact-hint" class="govuk-hint">
-          Select one option.
-        </div>
-        {{ govukRadios(model.contact) }}
-
-        <input type="hidden" name="taskName" id="taskName" value="{{ model.taskName }}" />
-        {% if model.person.email %}
-        <input type="hidden" name="email" id="email" value="{{ model.person.email }}" />
-        {% endif %}
-      </fieldset>
-
-      {{ govukButton({
-        text: "Send application",
-        name: "confirmSubmit" 
-      }) }}
-    </form>
-  </div>
-</div>
-
-{% endblock %}
